Part 086

+++ lib/models/detail.dart
@@ -3,15 +3,15 @@ import 'package:vehicles_app/models/procedure.dart';
 class Detail {
   int id = 0;
   Procedure procedure = Procedure(description: '', id: 0, price: 0);
-  double laborPrice = 0;
-  double sparePartsPrice = 0;
-  double totalPrice = 0;
+  int laborPrice = 0;
+  int sparePartsPrice = 0;
+  int totalPrice = 0;
   String? remarks = '';
 
   Detail({
     required this.id,
     required this.procedure,
-    required this.laborPrice,
+    required this.laborPrice, 
     required this.sparePartsPrice,
     required this.totalPrice,
     required this.remarks
