+++ lib/models/history.dart
@@ -8,9 +8,9 @@ class History {
   String? remarks = '';
   List<Detail> details = [];
   int detailsCount = 0;
-  double totalLabor = 0;
-  double totalSpareParts = 0;
-  double total = 0;
+  int totalLabor = 0;
+  int totalSpareParts = 0;
+  int total = 0;
 
   History({
     required this.id,
